+++ src/utilities.py
@@ -1,19 +1,7 @@
 from functools import partial
 import hashlib
-import sys
-import os
 from PyQt5.QtWidgets import QMessageBox
-import urllib3
-from constants import Constants, Signal, Database, ChecksumWhat
-
-
-def resource_path(relative_path):
-    """Get absolute path to resource, works for dev and for PyInstaller."""
-    try:
-        base_path = sys._MEIPASS
-    except Exception:
-        base_path = os.path.abspath(".")
-    return os.path.join(base_path, relative_path)
+from constants import Constants, Signal
 
 
 def uncheck_and_emit(button):
@@ -72,46 +60,15 @@ def pop_up(cls, title, text,
     return msg
 
 
-def is_mac_os():
-    """Return True if running OS is Mac."""
-    return sys.platform == 'darwin'
+def checksum_ok(data, reference_hash_code):
+    """Check whether the checksum of the 'data' argument is correct.
 
-
-def get_cacert_file():
-    """Return the path to the cacert.pem file."""
-    if hasattr(sys, "_MEIPASS"):
-        ca_certs = os.path.join(sys._MEIPASS, 'cacert.pem')
-    else:
-        ca_certs = 'cacert.pem'
-    return ca_certs
-
-
-def get_pool_manager():
-    """Return a urllib3.PoolManager object."""
-    return urllib3.PoolManager(ca_certs=get_cacert_file())
-
-
-def checksum_ok(data, what):
-    """Check whether the checksum of the 'data' argument is correct."""
+    Expects a sha256 code as argument."""
+    if reference_hash_code is None:
+        raise Exception("ERROR: Invalid hash code.")
     code = hashlib.sha256()
     code.update(data)
-    if what is ChecksumWhat.FOLDER:
-        n = 0
-    elif what is ChecksumWhat.DB:
-        n = 1
-    else:
-        raise ValueError("Wrong entry name.")
-    try:
-        # The downloaded file is a csv file with columns (last version == last line):
-        # data.zip_SHA256 | db.csv_SHA256 | Version | Creation_date
-        reference = get_pool_manager().request(
-            'GET',
-            Database.LINK_REF,
-            timeout=4.0
-        ).data.decode("utf-8").splitlines()[-1].split(Database.DELIMITER)[n]
-    except Exception:
-        raise
-    return code.hexdigest() == reference
+    return code.hexdigest() == reference_hash_code
 
 
 def connect_events_to_func(events_to_connect, fun_to_connect, fun_args):